+++ src/components/EmailEditor/components/Addons.vue
@@ -32,6 +32,21 @@
 							<!--<p class="ma-0">{{ item.description }}</p>-->
 						</v-col>
 
+						<v-col
+							v-if="item.state === 'disabled'"
+							class="align-content-right"
+							cols="6"
+						>
+							<v-text-field
+								:value="getDisabledReason(name)"
+								hide-details="true"
+								dense
+								outlined
+								label="disabledReason"
+								@input="setDisabledReason(name, $event)"
+							/>
+						</v-col>
+
 						<v-col class="align-self-center" cols="2">
 							<v-card flat class="d-flex justify-end align-center">
 								<v-btn
@@ -110,10 +125,20 @@ export default {
 		},
 	},
 	methods: {
-		...mapMutations([ "updateAddonState" ]),
+		...mapMutations([
+			"updateAddonState",
+			"updateAddonDisabledState",
+		]),
 		...mapActions({
 			updateEditorSettings: "updateEditorSettings",
 		}),
+		getDisabledReason(name) {
+			return this.$store.state.editorConfig.settings.addons[name].disabledReason;
+		},
+		setDisabledReason(name, value) {
+			this.updateAddonDisabledState({ name, value });
+			this.updateEditorSettings();
+		},
 	},
 
 };